backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
pandas
numpy
scikit-learn
joblib
openai
passlib[bcrypt]
PyJWT
